[IMP] Moved update_current_case out of util.py into blueprints namespace
